pyproject.toml
[tool.poetry]
name = "ultrasound-fetal-images"
version = "0.1.0"
description = ""
authors = ["Aleksander Zamojski <zamojski96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# --------- pytorch --------- #
torch = "^1.13.1"
torchvision = "^0.14.1"
pytorch-lightning = "1.8.6"
torchmetrics = "^0.11.0"
# --------- hydra --------- #
hydra-core = "1.3.1"
hydra-colorlog = "^1.2.0"
hydra-optuna-sweeper = "^1.2.0"
# --------- data --------- #
pandas = "^1.5.2"
numpy = "^1.24.1"
scikit-learn = "^1.2.0"
jupyter = "^1.0.0"
matplotlib = "^3.6.2"
# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# --------- others --------- #
pyrootutils = "^1.0.4"     # standardizing the project root setup
pre-commit = "^2.21.0"     # hooks for applying linters on commit
rich = "^12.6.0"           # beautiful text formatting in terminal
pytest = "^7.2.0"          # tests
pytest-cov = "^4.0.0"
sh = { version = "^1.14.3", markers = "sys_platform == 'linux' or sys_platform == 'darwin'" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--color=yes",
    "--durations=0",
    "--strict-markers",
    "--doctest-modules",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
log_cli = "True"
markers = [
    "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
